requirements.txt
APScheduler==3.7.0
arrow==0.17.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bfg==0.1.0
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.6
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-bigquery==2.7.0
google-cloud-core==1.5.0
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
h2==2.6.2
hpack==3.0.0
html5lib==1.1
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
lxml==4.6.2
numpy==1.19.5
pandas==1.1.5
paramiko==2.7.2
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
selenium==3.141.0
six==1.15.0
soupsieve==2.1
tzlocal==2.1
urllib3==1.26.2
webencodings==0.5.1
xlrd==2.0.1
